# Improve performance of Pipe.Reader.AsStream()