@davidfowl 2019-03-08
We really need to port our performance tests from Kestrel to dotnet/performance. One test I'd like to see if true async reads to see if it negatively impacts that scenario. Something like this benchmark https://github.com/aspnet/AspNetCore/blob/c79b2413ccd339ad12e39477371d128da1de2ec9/src/Servers/Kestrel/perf/Kestrel.Performance/PipeThroughputBenchmark.cs#L12