@stephentoub 2019-03-08
Thanks. I'll try that and share results.