@adamsitnik 2019-03-08
> Thanks. I'll try that and share results.

@stephentoub can you also add these benchmarks to dotnet/performance?