@stephentoub 2019-04-06
Going to close this until I have time to test/investigate the multi-threaded cases.